requirements.txt
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
motor==2.5.0
numpy
opencv-python
pillow
psutil
py-tgcalls==0.9.7
pykeyboard
pyrofork
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
uvloop
yt-dlp
youtube-search
youtube-search-python
httpx[http2]==0.27.2
bs4
wget
blackpink
